HDX-9781 Move humanitarian_needs VAT
