+++ src/hapi_schema/db_humanitarian_needs.py
@@ -3,7 +3,6 @@
 from datetime import datetime
 
 from sqlalchemy import (
-    Boolean,
     DateTime,
     Enum,
     ForeignKey,
@@ -126,39 +125,3 @@ view_params_humanitarian_needs = ViewParams(
         )
     ),
 )
-
-
-class DBHumanitarianNeedsVAT(Base):
-    __tablename__ = "humanitarian_needs_vat"
-    resource_hdx_id: Mapped[str] = mapped_column(String(36))
-    admin2_ref: Mapped[int] = mapped_column(
-        Integer, index=True, primary_key=True
-    )
-    gender: Mapped[str] = mapped_column(String(11), primary_key=True)
-    age_range: Mapped[str] = mapped_column(String(32), primary_key=True)
-    min_age: Mapped[int] = mapped_column(Integer, index=True)
-    max_age: Mapped[int] = mapped_column(Integer, index=True)
-    sector_code: Mapped[str] = mapped_column(String(32), primary_key=True)
-    population_group: Mapped[str] = mapped_column(String(14), primary_key=True)
-    population_status: Mapped[str] = mapped_column(
-        String(10), primary_key=True
-    )
-    disabled_marker: Mapped[str] = mapped_column(String(3), primary_key=True)
-    population: Mapped[int] = mapped_column(Integer, primary_key=True)
-    reference_period_start: Mapped[datetime] = mapped_column(
-        DateTime, primary_key=True
-    )
-    reference_period_end: Mapped[datetime] = mapped_column(
-        DateTime, index=True
-    )
-    sector_name: Mapped[str] = mapped_column(String(512))
-    location_code: Mapped[str] = mapped_column(String(128))
-    location_name: Mapped[str] = mapped_column(String(512), index=True)
-    location_ref: Mapped[int] = mapped_column(Integer)
-    admin1_code: Mapped[str] = mapped_column(String(128))
-    admin1_name: Mapped[str] = mapped_column(String(512))
-    admin1_is_unspecified: Mapped[bool] = mapped_column(Boolean)
-    admin2_code: Mapped[str] = mapped_column(String(128), index=True)
-    admin2_name: Mapped[str] = mapped_column(String(512), index=True)
-    admin2_is_unspecified: Mapped[bool] = mapped_column(Boolean)
-    admin1_ref: Mapped[int] = mapped_column(Integer)
