+++ src/hapi_schema/db_views_as_tables.py
@@ -104,3 +104,37 @@ class DBFundingVAT(Base):
     )
     location_code: Mapped[str] = mapped_column(String(128), index=True)
     location_name: Mapped[str] = mapped_column(String(512), index=True)
+
+
+class DBHumanitarianNeedsVAT(Base):
+    __tablename__ = "humanitarian_needs_vat"
+    resource_hdx_id: Mapped[str] = mapped_column(String(36))
+    admin2_ref: Mapped[int] = mapped_column(Integer, primary_key=True)
+    gender: Mapped[str] = mapped_column(String(11), primary_key=True)
+    age_range: Mapped[str] = mapped_column(String(32), primary_key=True)
+    min_age: Mapped[int] = mapped_column(Integer, nullable=True, index=True)
+    max_age: Mapped[int] = mapped_column(Integer, nullable=True, index=True)
+    sector_code: Mapped[str] = mapped_column(String(32))
+    population_group: Mapped[str] = mapped_column(String(14), primary_key=True)
+    population_status: Mapped[str] = mapped_column(
+        String(10), primary_key=True
+    )
+    disabled_marker: Mapped[str] = mapped_column(String(3), primary_key=True)
+    population: Mapped[int] = mapped_column(Integer, primary_key=True)
+    reference_period_start: Mapped[datetime] = mapped_column(
+        DateTime, primary_key=True
+    )
+    reference_period_end: Mapped[datetime] = mapped_column(
+        DateTime, index=True
+    )
+    sector_name: Mapped[str] = mapped_column(String(512))
+    location_code: Mapped[str] = mapped_column(String(128), index=True)
+    location_name: Mapped[str] = mapped_column(String(512), index=True)
+    location_ref: Mapped[int] = mapped_column(Integer)
+    admin1_code: Mapped[str] = mapped_column(String(128), index=True)
+    admin1_name: Mapped[str] = mapped_column(String(512), index=True)
+    admin1_is_unspecified: Mapped[bool] = mapped_column(Boolean)
+    admin2_code: Mapped[str] = mapped_column(String(128), index=True)
+    admin2_name: Mapped[str] = mapped_column(String(512), index=True)
+    admin2_is_unspecified: Mapped[bool] = mapped_column(Boolean)
+    admin1_ref: Mapped[int] = mapped_column(Integer)
